test: add diagnostic and port format helper coverage
(cherry picked from commit f95c09b6f2cd6dda90077dc62f42812c23a7f9a9)
