+++ src/infra/ports.test.ts
@@ -7,16 +7,8 @@ const runCommandWithTimeoutMock = vi.hoisted(() => vi.fn());
 vi.mock("../process/exec.js", () => ({
   runCommandWithTimeout: (...args: unknown[]) => runCommandWithTimeoutMock(...args),
 }));
-import { formatPortListener } from "./ports-format.js";
 import { inspectPortUsage } from "./ports-inspect.js";
-import {
-  buildPortHints,
-  classifyPortListener,
-  ensurePortAvailable,
-  formatPortDiagnostics,
-  handlePortError,
-  PortInUseError,
-} from "./ports.js";
+import { ensurePortAvailable, handlePortError, PortInUseError } from "./ports.js";
 
 const describeUnix = process.platform === "win32" ? describe.skip : describe;
 
@@ -61,78 +53,6 @@ describe("ports helpers", () => {
 
     const messages = runtime.error.mock.calls.map((call) => stripAnsi(String(call[0] ?? "")));
     expect(messages.join("\n")).toContain("another RemoteClaw instance is already running");
-  });
-
-  it("classifies port listeners across gateway, ssh, and unknown cases", () => {
-    const cases = [
-      {
-        listener: { commandLine: "ssh -N -L 18789:127.0.0.1:18789 user@host" },
-        expected: "ssh",
-      },
-      {
-        listener: { command: "ssh" },
-        expected: "ssh",
-      },
-      {
-        listener: { commandLine: "node /Users/me/Projects/remoteclaw/dist/entry.js gateway" },
-        expected: "gateway",
-      },
-      {
-        listener: { commandLine: "python -m http.server 18789" },
-        expected: "unknown",
-      },
-    ] as const;
-
-    for (const testCase of cases) {
-      expect(
-        classifyPortListener(testCase.listener, 18789),
-        JSON.stringify(testCase.listener),
-      ).toBe(testCase.expected);
-    }
-  });
-
-  it("builds ordered hints for mixed listener kinds and multiple listeners", () => {
-    expect(
-      buildPortHints(
-        [
-          { commandLine: "node dist/index.js remoteclaw gateway" },
-          { commandLine: "ssh -N -L 18789:127.0.0.1:18789" },
-          { commandLine: "python -m http.server 18789" },
-        ],
-        18789,
-      ),
-    ).toEqual([
-      expect.stringContaining("Gateway already running locally."),
-      "SSH tunnel already bound to this port. Close the tunnel or use a different local port in -L.",
-      "Another process is listening on this port.",
-      expect.stringContaining("Multiple listeners detected"),
-    ]);
-    expect(buildPortHints([], 18789)).toEqual([]);
-  });
-
-  it("formats port listeners and diagnostics for free and busy ports", () => {
-    expect(formatPortListener({ command: "ssh", address: "127.0.0.1:18789" })).toBe(
-      "pid ?: ssh (127.0.0.1:18789)",
-    );
-
-    expect(
-      formatPortDiagnostics({
-        port: 18789,
-        status: "free",
-        listeners: [],
-        hints: [],
-      }),
-    ).toEqual(["Port 18789 is free."]);
-
-    const lines = formatPortDiagnostics({
-      port: 18789,
-      status: "busy",
-      listeners: [{ pid: 123, user: "alice", commandLine: "ssh -N -L 18789:127.0.0.1:18789" }],
-      hints: buildPortHints([{ pid: 123, commandLine: "ssh -N -L 18789:127.0.0.1:18789" }], 18789),
-    });
-    expect(lines[0]).toContain("Port 18789 is already in use");
-    expect(lines).toContain("- pid 123 alice: ssh -N -L 18789:127.0.0.1:18789");
-    expect(lines.some((line) => line.includes("SSH tunnel"))).toBe(true);
   });
 });
